organize the data import module
