+++ database/get_acc_latlng.py
@@ -1,10 +1,8 @@
 import sys
 sys.path.append('../')
 
-from consts import PG_HOST, PG_DB, PG_USER, PG_PASSWD
 from consts import MG_HOST, MG_PORT
 
-import psycopg2
 from pymongo import MongoClient
 
 client = MongoClient(MG_HOST, MG_PORT)
@@ -14,41 +12,6 @@ lrs_col = db.lrs
 
 from shapely.wkt import loads as wkt_loads
 from shapely.geometry import LineString
-
-def cnty_rte2linestring(rid):
-    """Consolidate accident road dataset with the LRS system
-    :param rid: the road id of interest
-    """
-    rte_info = lrs_col.find_one({'rid':rid})
-    if rte_info:
-        return rte_info
-
-    # first reformat the rid
-    # move the first two id to the end
-    # 2530000024 -> 3000002425
-    new_rid = rid
-    conn = psycopg2.connect("dbname='%s' user='%s' host='%s' password='%s'"%(PG_DB, PG_USER, PG_HOST, PG_PASSWD) )
-    sql = "select begmp1, endmp1, st_astext(geom) from ncdot84 where rte_id = '%s' order by begmp1;"%(new_rid)
-    cur = conn.cursor()
-    cur.execute(sql)
-    rows = cur.fetchall()
-    rlist = []
-    plist = []
-    for row in rows:
-        row_info = {
-                'begmp': float(row[0]),
-                'endmp': float(row[1]),
-                'linestring': row[2]
-                }
-
-        # construct the whole linestring
-        wkt_ls = wkt_loads( row[2] )
-        seg_plist = list( wkt_ls[0].coords )
-        plist.extend( seg_plist )
-        rlist.append(row_info)
-    rte_info = {'rid':rid, 'rlist': rlist, 'plist':plist}
-    lrs_col.insert(rte_info)
-    return rte_info
 
 def acc2latlng(caseno):
     """Calculate the latitude/longitude and update into mongodb
