+++ database/import_additional.py
@@ -20,6 +20,8 @@ ylist = ['09','10','11','12','13']
 hsis_data_folder = "../../TTI/data/HSIS"
 
 def update_acc():
+    """ add the alchol usage information
+    """
     name = 'acc'
     for y in ylist:
         f = SAS7BDAT('%s/NC/nc%s%s.sas7bdat'%(hsis_data_folder,y,name))
@@ -36,6 +38,8 @@ def update_acc():
                 )
 
 def update_acc2():
+    """add number of lanes and the location type
+    """
     name = 'acc'
     for y in ylist:
         f = SAS7BDAT('%s/NC/nc%s%s.sas7bdat'%(hsis_data_folder,y,name))
@@ -71,6 +75,8 @@ def update_road():
             )
 
 def update_vehicle():
+    """ add the injury, achohol usage, and vision obscurity for each vehicle
+    """
     print 'update vehicle'
     name = 'veh'
     field_list = ['alcflag','vision','num_k','num_a','num_b','num_c']
@@ -89,66 +95,9 @@ def update_vehicle():
                 update_dict[field_list[j]] = data_list_list[j][i]
             veh_col.update( filter_dict, {'$set':update_dict} )
 
-
-def consolidate_injury():
-    # get all caseno
-    caseno_list = []
-    for acc in acc_col.find({'num_k':{'$exists':False}}):
-        caseno_list.append( acc['caseno'] )
-
-    # find all vehicle, get the number and consolidate
-    for caseno in caseno_list:
-        try:
-            vehs = veh_col.find({'caseno':caseno})
-            nk = 0
-            na = 0
-            nb = 0
-            nc = 0
-            for veh in vehs:
-
-                nk += veh['num_k']
-                na += veh['num_a']
-                nb += veh['num_b']
-                nc += veh['num_c']
-
-            acc_col.update(
-                {'caseno':caseno},
-                {'$set':{
-                    'num_k':nk,
-                    'num_a':na,
-                    'num_b':nb,
-                    'num_c':nc,
-                    }})
-        except Exception  as e:
-            print e
-
-def consolidate_colision_type():
-    caseno_list = []
-    for acc in acc_col.find({'events':{'$exists':False}}):
-        caseno_list.append( acc['caseno'] )
-
-    # find all vehicle, get the number and consolidate
-    for caseno in caseno_list:
-        try:
-            vehs = veh_col.find({'caseno':caseno})
-            events = []
-            for veh in vehs:
-                for i in range(1,5):
-                    if not np.isnan(veh['event'+str(i)]):
-                        events.append(veh['event'+str(i)])
-            events = list(set(events))
-            #print events
-            #break
-            acc_col.update(
-                {'caseno':caseno},
-                {'$set':{'events':events}})
-        except Exception  as e:
-            print e
-
 if __name__ == "__main__":
     #update_acc()
     #update_acc2()
     #update_road()
     #update_vehicle()
-    #consolidate_injury()
-    consolidate_colision_type()
+    pass
